+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - [#2102](https://github.com/slimphp/Slim/pull/2102) Remove container from router
 - [#2124](https://github.com/slimphp/Slim/pull/2124) Remove Slim\Exception\SlimException
 - [#2174](https://github.com/slimphp/Slim/pull/2174) Switch from Container-Interop to PSR-11
+- [#2290](https://github.com/slimphp/Slim/pull/2290) Removed container. Set your own using `App::setContainer()`
 
 ### Fixed